Resolve type error on YAML
