+++ vercel/src/routes/ascii-preview/Index.svelte
@@ -2,7 +2,7 @@
   import Nav from '../../lib/Nav.svelte';
   import AsciiPreview from '../../lib/AsciiPreview.svelte';
   import data from '../../../../languages.yaml';
-  import type { Languages } from 'src/types';
+  import type { Languages } from '../../../../languages.yaml';
 
   const languages = Object.entries(data as Languages).map(
     ([name, { ascii, colors }]) => ({
